Fix -Wconversion warnings

+++ example_callbacks/http_compression/http_compression_callback.c
@@ -2,9 +2,9 @@
  *
  * Response body compression callback function for Ulfius Framework
  *
- * Copyright 2020 Nicolas Mora <mail@babelouest.org>
+ * Copyright 2020-2022 Nicolas Mora <mail@babelouest.org>
  *
- * Version 20201213
+ * Version 20221024
  *
  * Compress the response body using `deflate` or `gzip` depending on the request header `Accept-Encoding` and the callback configuration.
  * The rest of the response, status, headers, cookies won't change.
@@ -125,7 +125,7 @@ int callback_http_compression (const struct _u_request * request, struct _u_resp
           } while (U_CALLBACK_IGNORE == ret && defstream.avail_out == 0);
 
           if (ret == U_CALLBACK_IGNORE) {
-            ulfius_set_binary_body_response(response, response->status, (const char *)data_zip, defstream.total_out);
+            ulfius_set_binary_body_response(response, (unsigned int)response->status, (const char *)data_zip, defstream.total_out);
             u_map_put(response->map_header, U_CONTENT_HEADER, compress_mode==U_COMPRESS_GZIP?U_ACCEPT_GZIP:U_ACCEPT_DEFLATE);
           }
           deflateEnd(&defstream);
